+++ src/main/kotlin/com/github/mrdolch/externaljavaformatter/PersistConfigurationService.kt
@@ -17,8 +17,8 @@ class PersistConfigurationService : PersistentStateComponent<PersistConfiguratio
   }
 
   class Configuration {
-    var enabled: Boolean? = false
-    var useStandardIn: Boolean? = true
+    var enabled: Boolean = false
+    var useStandardIn: Boolean = true
     var classPath: String? = "configurable-java-format-2026.1.0-all-deps.jar"
     var mainClass: String? = "com.google.googlejavaformat.java.Main"
     var arguments: String? = "--width=120 --assume-filename {} -"